+++ src/main/scala/de/sciss/synth/GEOps.scala
@@ -13,7 +13,6 @@
 
 package de.sciss.synth
 
-import de.sciss.optional.Optional
 import de.sciss.synth.ugen.{BinaryOpUGen, ChannelProxy, Clip, Constant, Flatten, Fold, Impulse, LinExp, LinLin, MulAdd, Poll, UnaryOpUGen, Wrap}
 
 object GEOps {
@@ -21,8 +20,8 @@ object GEOps {
     g.rate.getOrElse(throw new UnsupportedOperationException(s"`$name` input rate must be defined"))
 }
 final class GEOps(val `this`: GE) extends AnyVal { me =>
-  import me.{`this` => g}
   import GEOps.getRate
+  import me.{`this` => g}
 
   /** Creates a proxy that represents a specific output channel of the element.
     *
@@ -46,19 +45,19 @@ final class GEOps(val `this`: GE) extends AnyVal { me =>
     *                   interpreted as a frequency value and an `Impulse` generator of that frequency
     *                   is used instead.
     * @param   label    a string to print along with the values, in order to identify
-    *                   different polls. Using the special label `"#auto"` (default) will generated
+    *                   different polls. Using the special label `"$auto"` (default) will generated
     *                   automatic useful labels using information from the polled graph element
     * @param   trigID   if greater then 0, a `"/tr"` OSC message is sent back to the client
     *                   (similar to `SendTrig`)
     *
     * @see  [[de.sciss.synth.ugen.Poll]]
     */
-  def poll(trig: GE = 10, label: Optional[String] = None, trigID: GE = -1): Poll = {
+  def poll(trig: GE = 10, label: String = "$auto", trigID: GE = -1): Poll = {
     val trig1 = trig match {
       case Constant(freq) => Impulse((g.rate getOrElse audio) max control, freq, 0) // XXX good? or throw an error? should have a maxRate?
       case other          => other
     }
-    Poll(trig1.rate, trig1, g, label.getOrElse {
+    val label1 = if (label != "$auto") label else {
       val str = g.toString
       val i   = str.indexOf('(')
       if (i >= 0) str.substring(0, i)
@@ -67,7 +66,8 @@ final class GEOps(val `this`: GE) extends AnyVal { me =>
         if (j >= 0) str.substring(0, j)
         else str
       }
-    }, trigID)
+    }
+    Poll(trig1.rate, trig = trig1, in = g, label = label1, trigID = trigID)
   }
 
   import UnaryOpUGen._