+++ src/test/scala/de/sciss/synth/GraphFunctionTest.scala
@@ -1,0 +1,11 @@
+package de.sciss.synth
+
+trait GraphFunctionTest {
+  import Ops._
+  import ugen._
+
+  val x = graph {
+    val f = LFSaw.kr(0.4).madd(24, LFSaw.kr(List(8.0, 7.23)).madd(3, 80)).midicps
+    CombN.ar(SinOsc.ar(f)*0.04, 0.2, 0.2, 4)
+  }
+}